pro_browser/node/requirements.txt
playwright
aiortc
python-socketio[asyncio_client]
numpy
Pillow
opencv-python-headless
